services/report-service/pyproject.toml
[project]
name = "report-service"
version = "1.0.0"
description = "CaisseFlow Pro Report Service — PDF, Excel, CSV report generation"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.30",
  "pyodbc>=5.1.0",
  "aioodbc>=0.5.0",
  "celery[redis]>=5.4.0",
  "redis>=5.0.0",
  "reportlab>=4.2.0",
  "weasyprint>=62.0",
  "openpyxl>=3.1.5",
  "python-jose[cryptography]>=3.3.0",
  "jinja2>=3.1.4",
  "httpx>=0.27.0",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
